requirements.txt
argcomplete==2.0.0
asgiref==3.6.0
autopep8==2.0.1
Beaker==1.12.1
beautifulsoup4==4.12.2
blivet==3.7.1
blivet-gui==2.4.1
Brlapi==0.8.4
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.6
distro==1.8.0
dj-database-url==1.0.0
Django==4.1.5
djangorestframework==3.14.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.8.2
fros==1.1
gpg==1.17.1
gunicorn==20.0.4
humanize==3.13.1
idna==3.4
jaraco.classes==3.2.3
langtable==0.0.62
libcomps==0.1.18
lxml==4.9.2
Mako==1.2.3
MarkupSafe==2.1.2
mod-wsgi==4.9.4
more-itertools==8.13.0
nftables==0.1
numpy==1.24.4
olefile==0.46
packaging==23.0
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.4.0
pipenv==2022.10.25
platformdirs==2.6.0
ply==3.11
productmd==1.35
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGObject==3.44.1
pykickstart==3.47
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyparted==3.12.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-meh==0.51
pytz==2022.7
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.18.1
selinux==3.5
sepolicy==3.5
setools==4.4.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.5.1
soupsieve==2.4.1
sqlparse==0.4.3
systemd-python==235
Tempita==0.5.2
urllib3==1.26.16
virtualenv==20.21.1
virtualenv-clone==0.5.7
